Create sampl.subquery.sql
The query above mentioned is to find the salesperson with the highest total sales using subqueries. Subqueries are queries within queries. These can be used for filtering, creating derived tables, or complex aggregations.

In this case, the outer query uses a subquery to identify the salesperson with the highest sales.